migration/migrationutils: add package to store utilities
We have some utilities that are shared between startup migrations and
migrations. We're going to be adding more. It'd be a shame to too tightly
couple some of these tools to the migration they were built for.

Release note: None
