+++ pkg/BUILD.bazel
@@ -148,6 +148,7 @@ ALL_TESTS = [
     "//pkg/migration/migrationcluster:migrationcluster_test",
     "//pkg/migration/migrationmanager:migrationmanager_test",
     "//pkg/migration/migrations:migrations_test",
+    "//pkg/migration/migrationutils:migrationschemautils_test",
     "//pkg/roachpb:roachpb_test",
     "//pkg/roachpb:string_test",
     "//pkg/rpc/nodedialer:nodedialer_test",
