+++ pkg/startupmigrations/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
         "//pkg/jobs",
         "//pkg/keys",
         "//pkg/kv",
+        "//pkg/migration/migrationutils",
         "//pkg/roachpb:with-mocks",
         "//pkg/security",
         "//pkg/settings/cluster",
@@ -54,10 +55,6 @@ go_test(
         "//pkg/security/securitytest",
         "//pkg/server",
         "//pkg/sql",
-        "//pkg/sql/catalog/catalogkeys",
-        "//pkg/sql/catalog/descpb",
-        "//pkg/sql/catalog/systemschema",
-        "//pkg/sql/catalog/tabledesc",
         "//pkg/startupmigrations/leasemanager",
         "//pkg/testutils",
         "//pkg/testutils/serverutils",
